feat: add controller and service

+++ domain/src/main/kotlin/com/jvmhater/moduticket/model/query/Page.kt
@@ -1,3 +1,3 @@
 package com.jvmhater.moduticket.model.query
 
-data class Page(val page: Int, val size: Int)
+data class Page(val page: Long, val size: Int)
